# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: 极简栈创造 AI 生成接口
  description: |
    涉及 AI 适配器层的对外接口规范，包括模型绑定、生成任务、Adapter 健康检查。
  version: 1.0.0
  contact:
    name: Core · 后端工程师
servers:
  - url: /api/v1
    description: 本地开发服务
paths:
...
  /projects/{projectId}/characters:
    parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      summary: 列出项目角色
      description: 返回指定项目下的所有角色卡片。
      operationId: listCharacters
      tags:
        - Characters
      responses:
        '200':
          description: 角色列表
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    $ref: '#/components/schemas/CharacterList'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      summary: 创建角色卡片
      description: 手动创建角色卡片。
      operationId: createCharacter
      tags:
        - Characters
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCharacterRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    $ref: '#/components/schemas/Character'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /projects/{projectId}/characters/auto-create:
    parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      summary: 根据大纲自动创建角色
      description: 根据已确认大纲中的 characters[] 批量创建角色卡片。
      operationId: autoCreateCharacters
      tags:
        - Characters
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AutoCreateCharactersRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Character'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /projects/{projectId}/characters/{characterId}:
    parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: characterId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      summary: 获取角色详情
      operationId: getCharacter
      tags:
        - Characters
      responses:
        '200':
          description: 角色详情
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    $ref: '#/components/schemas/Character'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      summary: 更新角色卡片
      description: 更新角色字段，自动触发版本快照。
      operationId: updateCharacter
      tags:
        - Characters
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCharacterRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    $ref: '#/components/schemas/Character'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: 删除角色
      operationId: deleteCharacter
      tags:
        - Characters
      responses:
        '204':
          description: 删除成功
        '404':
          $ref: '#/components/responses/NotFound'
  /projects/{projectId}/characters/{characterId}/generate-views:
    parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: characterId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      summary: 生成三视图
      description: 为角色生成正面、侧面、背面三视图。
      operationId: generateCharacterViews
      tags:
        - Characters
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateViewsRequest'
      responses:
        '200':
          description: 生成成功
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    $ref: '#/components/schemas/Character'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /projects/{projectId}/characters/{characterId}/generate-views/{viewId}/retry:
    parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: characterId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: viewId
        in: path
        required: true
        schema:
          type: string
          enum:
            - front
            - side
            - back
    post:
      summary: 重新生成单张视图
      description: 对指定视图（front/side/back）重新生成。
      operationId: retryCharacterView
      tags:
        - Characters
      responses:
        '200':
          description: 重新生成成功
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    $ref: '#/components/schemas/Character'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /projects/{projectId}/characters/{characterId}/confirm-views:
    parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: characterId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      summary: 确认三视图
      description: 确认三视图并锁定参考图，自动触发扩展参考图生成。
      operationId: confirmCharacterViews
      tags:
        - Characters
      responses:
        '200':
          description: 确认成功
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    $ref: '#/components/schemas/Character'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /projects/{projectId}/characters/{characterId}/generate-refs:
    parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: characterId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      summary: 生成扩展参考图
      description: 生成表情组（4 张）与场景站立图（2 张）。
      operationId: generateCharacterRefs
      tags:
        - Characters
      responses:
        '200':
          description: 生成成功
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    $ref: '#/components/schemas/Character'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /projects/{projectId}/characters/{characterId}/rollback:
    parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: characterId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      summary: 回滚到指定版本
      description: 基于历史版本快照恢复角色数据。
      operationId: rollbackCharacter
      tags:
        - Characters
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RollbackCharacterRequest'
      responses:
        '200':
          description: 回滚成功
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    $ref: '#/components/schemas/Character'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  schemas:
...
    CharacterRoleType:
      type: string
      enum:
        - protagonist
        - supporting
        - antagonist
      description: 角色定位
    CharacterStatus:
      type: string
      enum:
        - draft
        - confirmed
      description: 角色状态
    CharacterRefImage:
      type: object
      required:
        - view
        - url
        - seed
      properties:
        view:
          type: string
          description: 视图或参考图类型标识
        url:
          type: string
          format: uri
          description: 图片 URL
        seed:
          type: integer
          description: 生成使用的种子
    Character:
      type: object
      required:
        - id
        - project_id
        - name
        - role_type
        - status
        - ref_images
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        project_id:
          type: string
          format: uuid
        name:
          type: string
        role_type:
          $ref: '#/components/schemas/CharacterRoleType'
        episode_range:
          type: string
          nullable: true
          description: 登场集数范围
        appearance:
          type: string
          nullable: true
          description: 外貌描述
        costume:
          type: string
          nullable: true
          description: 服装方案
        expression:
          type: string
          nullable: true
          description: 表情特征
        signature_action:
          type: string
          nullable: true
          description: 标志性动作
        voice_description:
          type: string
          nullable: true
          description: 声线描述
        status:
          $ref: '#/components/schemas/CharacterStatus'
        ref_images:
          type: array
          items:
            $ref: '#/components/schemas/CharacterRefImage'
        ip_adapter_id:
          type: string
          nullable: true
          description: IP-Adapter 角色 Embedding ID
        lora_path:
          type: string
          nullable: true
          description: LoRA 模型路径
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CharacterList:
      type: object
      required:
        - total
        - characters
      properties:
        total:
          type: integer
        characters:
          type: array
          items:
            $ref: '#/components/schemas/Character'
    CreateCharacterRequest:
      type: object
      required:
        - name
        - role_type
      properties:
        name:
          type: string
          maxLength: 100
        role_type:
          $ref: '#/components/schemas/CharacterRoleType'
        episode_range:
          type: string
          maxLength: 50
          nullable: true
        appearance:
          type: string
          maxLength: 2000
          nullable: true
        costume:
          type: string
          maxLength: 2000
          nullable: true
        expression:
          type: string
          maxLength: 2000
          nullable: true
        signature_action:
          type: string
          maxLength: 500
          nullable: true
        voice_description:
          type: string
          maxLength: 500
          nullable: true
        status:
          $ref: '#/components/schemas/CharacterStatus'
    UpdateCharacterRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 100
        role_type:
          $ref: '#/components/schemas/CharacterRoleType'
        episode_range:
          type: string
          maxLength: 50
          nullable: true
        appearance:
          type: string
          maxLength: 2000
          nullable: true
        costume:
          type: string
          maxLength: 2000
          nullable: true
        expression:
          type: string
          maxLength: 2000
          nullable: true
        signature_action:
          type: string
          maxLength: 500
          nullable: true
        voice_description:
          type: string
          maxLength: 500
          nullable: true
        status:
          $ref: '#/components/schemas/CharacterStatus'
    AutoCreateCharactersRequest:
      type: object
      required:
        - characters
      properties:
        characters:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/CreateCharacterRequest'
    GenerateViewsRequest:
      type: object
      properties:
        seed:
          type: integer
          description: 固定种子，不传则根据角色 ID 与视图计算
        width:
          type: integer
        height:
          type: integer
        style_preset:
          type: string
    RollbackCharacterRequest:
      type: object
      required:
        - version_id
      properties:
        version_id:
          type: string
          pattern: '^v\d+$'
          description: 目标版本号
...
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: 统一错误码
          example: AI_TIMEOUT
        message:
          type: string
          description: 错误信息（不暴露堆栈）
        retryable:
          type: boolean
          description: 是否为可用性错误，可触发 fallback
  responses:
    BadRequest:
      description: 请求参数错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'